requirements.txt
absl-py==2.1.0
accelerate==0.20.3
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==5.0.1
attrs==24.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.3.1
cycler==0.12.1
datasets==3.1.0
diffusers==0.21.2
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.0
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.3.1
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
grpcio==1.68.1
httplib2==0.22.0
huggingface-hub==0.25.2
idna==3.10
imageio==2.36.1
importlib_metadata==8.5.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.1
psutil==6.1.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.5
scikit-image==0.25.0
scikit-learn==1.6.1
scipy==1.15.1
sentry-sdk==2.19.2
setproctitle==1.3.4
six==1.16.0
smmap==5.0.1
soupsieve==2.6
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2025.1.10
tokenizers==0.13.3
torch==2.5.1
torchaudio==2.5.1+cu121
torchvision==0.20.1
tqdm==4.67.1
transformers==4.25.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wandb==0.19.0
wcwidth==0.2.13
Werkzeug==3.1.3
xformers==0.0.28.post3
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
